requirements.txt
python-docx
deep-translator
tqdm
python-dotenv 